# docs/phase1_openapi.yaml
openapi: 3.0.3
info:
  title: 智享营销系统第 1 阶段 API
  version: 1.0.0
  description: >
    第 1 阶段 MVP 接口定义，覆盖后台管理、门店端、小程序端、微信支付、销售单微信分享收款。
    价格、库存、订单状态、客户身份、收款状态均由后端裁剪和流转，前端不得自行判断。
servers:
  - url: https://api.example.com
    description: 生产环境占位
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/miniapp/login:
    post:
      tags: [Miniapp]
      summary: 小程序登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: 微信登录 code
      responses:
        '200':
          description: 登录结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MiniappLoginResponse'
  /api/miniapp/products:
    get:
      tags: [Miniapp]
      summary: 小程序商品列表
      description: 后端按客户身份裁剪价格字段。普通客户不得返回批发价字段。
      security:
        - miniappAuth: []
      parameters:
        - name: storeId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: keyword
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 商品列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MiniappProduct'
  /api/miniapp/orders:
    post:
      tags: [Miniapp]
      summary: 创建小程序订单
      security:
        - miniappAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiniappOrderCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MiniappOrder'
  /api/miniapp/orders/{orderNo}:
    get:
      tags: [Miniapp]
      summary: 小程序订单详情
      security:
        - miniappAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MiniappOrder'
...
  /api/share/collections/{token}:
    get:
      tags: [ShareCollection]
      summary: 客户查看分享收款单
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 分享收款详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShareCollectionDetail'
  /api/share/collections/{token}/pay:
    post:
      tags: [ShareCollection]
      summary: 客户从分享页发起支付
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 预支付参数
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WechatPrepayResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    miniappAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderNo:
      name: orderNo
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "0"
        message:
          type: string
          example: 成功
        traceId:
          type: string
          example: 01HZTRACEID
...
    MiniappProduct:
      type: object
      description: 普通客户响应中不得包含 wholesalePrice 字段。
      properties:
        skuId:
          type: integer
          format: int64
        name:
          type: string
        skuName:
          type: string
        image:
          type: string
        price:
          type: number
          description: 当前客户实际可见价格
        retailPrice:
          type: number
        wholesalePrice:
          type: number
          nullable: true
          description: 仅批发客户可返回
        priceType:
          type: string
          enum: [RETAIL, WHOLESALE, MINIAPP]
        availableQty:
          type: integer
...
    MiniappOrderCreateRequest:
      type: object
      required: [storeId, fulfillmentType, items]
      properties:
        storeId:
          type: integer
          format: int64
        fulfillmentType:
          type: string
          enum: [DELIVERY, PICKUP]
        receiverName:
          type: string
        receiverMobile:
          type: string
        receiverAddress:
          type: string
        remark:
          type: string
        idempotencyKey:
          type: string
        items:
          type: array
          items:
            type: object
            required: [skuId, qty]
            properties:
              skuId:
                type: integer
                format: int64
              qty:
                type: integer
    MiniappOrder:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          type: string
        payStatus:
          type: string
        payableAmount:
          type: number
        expireAt:
          type: string
          format: date-time
...
    SaleBillItem:
      type: object
      properties:
        skuId:
          type: integer
          format: int64
        skuName:
          type: string
        boxQty:
          type: integer
        bottleQty:
          type: integer
        totalBottleQty:
          type: integer
        unitPrice:
          type: number
        subtotalAmount:
          type: number
...
    MiniappLoginResponse:
      type: object
      properties:
        token:
          type: string
        memberId:
          type: integer
          format: int64
        customerType:
          type: string
          enum: [RETAIL, WHOLESALE]
...
    WechatPrepayResponse:
      type: object
      properties:
        payNo:
          type: string
        appId:
          type: string
        timeStamp:
          type: string
        nonceStr:
          type: string
        package:
          type: string
        signType:
          type: string
          example: RSA
        paySign:
          type: string
...
    ShareCollectionDetail:
      type: object
      properties:
        linkNo:
          type: string
        storeName:
          type: string
        sourceNo:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
        expireAt:
          type: string
          format: date-time
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleBillItem'
...